Allow alternative `push` commands (#142)
Closes #140.

To push chart packages to ChartMuseum registries, one needs to run `helm
cm-push` instead of the default `helm push`.

The present change simply consists in adding an optional `push_cmd`
string argument to the `helm_chart` macro definition and propagate it
down to the registrar.

I've verified the change manually by overriding the `rules_helm`
repository using the present branch, but didn't find a simple way to
test it yet.