Fix dependency handling (#199)
## What?
Fixes a couple of problems with how dependencies are handled in charts

* Update the Chart.yaml of the main chart with the dependencies if they
are not already added manually to the Chart.yaml.
* Correctly use the name and version of the dependency when copying the
dependency tarball to the staging folder before the chart is packaged.
Prior to this fix the of the Bazel target was used and that causes the
bug that is reported in:
  *  #91
* Update tests. They were not being run due to missing `Test` prefix on
the name of the test.

Fixes #91